requirements.txt
# Core
numpy>=1.24.0
torch>=2.0.0
torchaudio>=2.0.0

# Transcription
faster-whisper>=1.0.0

# Shot detection
opencv-python>=4.8.0
scenedetect>=0.6.0

# Speaker diarization (optional)
pyannote.audio>=3.1.0

# LLM
openai>=1.0.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0

# Video processing
# ffmpeg and ffprobe must be installed system-wide
